Adding source, header, and Make files
These contain the implementation for the network.
Functionally, the network works in terms of
the feed forward mechanism and outputting a
probable answer.

The back propagation method technically  works as
well but does not seem to be able to actually
train the network successfully. The calculated
cost and it's derivative with respect to the
network weights and biases are off, so this needs
further debugging.

README.md changes are also included in this commit.