Updated Makefile to check training files
Updated the Makefile to include warning messages if
either of the training or testing data .csv files
are missing from the the ./data directory.

I defined the names of the files and the data directory
as variables, in case I reuse this project in the future
with different data sets or data formats.