Including Compiler Optimizations in Build Process:
The updated matrix struct and squite of functions is quite
slow. Using it for the back propagation algorithm means
that a lot more looping and operations are happening, not
to mention the excessive memory allocations. This change
is just an extra effort to speed up the execution time,
since the overall build time is very quick.