Adding Timer for Batch Learning:
I was curious to see how long each traing batch
was taking, since I felt that it was taking
longer the more batches had been processed.

This is true, which I believe still has to do
with memory allocation, but adding the timer
should be a reasonable representation of the
overall performance of the training.