Rewrote Back Propagation Implementation:
It turns out I was really off in terms of the
calculations I was making and how I was applying
the cost.

I fixed this, by more closely following the
process described in 3Blue1Brown's video series.
It also makes a lot more sense to me now. Currently
the model is consistently getting %11.3 accuracy,
never below ~%10, which means its learning something
but it's not super accurate yet.