Fixes to Back Propagation Implementation:
The code is still formatted in a debug style, but the
overall function of the back propagation seems to be
working. The changes I made earlier to the matrix
library make the execution of the algorithm across
even a thousand test samples extremely slow. To move
forward, I'll need to switch the matrix code back to
the old implementation style, where a pointer to the
'resulting' matrix is passed into functions. I didn't
want to deal managing the points as much, but I guess
I'm paying for that now.

Also updated the warning messages for missing data in
the Makefile.

Signed-off-by: mrandronov <mrandronov88@gmail.com>